refactor(analytics): extract ChampionProComputeService from the god-file (P10.1)
Continues the P10.1 decomposition by lifting the pro / high-elo domain out of
ChampionAnalyticsComputeService (+ its .Stats.cs partial) into a focused
ChampionProComputeService, mirroring the win-rate (#100) and builds (#102)
extractions.

- New IChampionProComputeService + ChampionProComputeService own the moved-verbatim
  ComputeProBuildsAsync, ComputeProChampionPlayrateAsync, ComputeProRosterAsync,
  ComputeProBuildsFromStatsAsync, and ComputeProChampionPlayrateFromStatsAsync.
  The new service takes (TranscendenceContext, IOptions<ChampionAnalyticsComputeOptions>)
  -- it reads ProBuildMaxParticipantRows but never logs, so no logger (matching the
  win-rate service). Reuses the shared ChampionBuildPathBuilder / AnalyticsScopeMath
  / AnalyticsSnapshotSerialization helpers (nothing duplicated).
- NormalizeProScope (internal static) moved with the pro domain; every external static
  call site (3 in ChampionAnalyticsService) rewired to ChampionProComputeService.
  NormalizeProScope(...). The moved *FromStats methods resolve it in-class.
- ChampionAnalyticsComputeService is now matchups-only: its ctor trimmed to (context)
  since matchups use neither options nor logger; class/interface docs refreshed.
- Rewired the orchestrator (ChampionAnalyticsService) to inject + route the pro
  *FromStats / roster / NormalizeProScope calls to the pro service (matchups stay on
  the compute service). PrecomputedAnalyticsRefresher now injects the pro service for
  its pro-surface refresh and drops the unused matchups-compute dependency (it builds
  matchup atoms directly in SQL). DI registers the pro service alongside the others.
- Tests retargeted: per-service CreateProService/ProService helpers, a
  Mock<IChampionProComputeService> added to the orchestrator Harness, and every
  PrecomputedAnalyticsRefresher construction updated for the new (context, build, pro,
  logger) shape. No asserted values changed.

Behavior-preserving, gated by the analytics tests (raw + raw-vs-stats pro-surface
equivalence): Service.Core 183/183, WebAPI 40/40.
