+++ tests/Transcendence.Service.Core.Tests/ChampionMatchupStatsEquivalenceTests.cs
@@ -27,7 +27,7 @@ public class ChampionMatchupStatsEquivalenceTests
     public async Task Matchups_StatsPath_EqualsRawCompute_AcrossRankScopes()
     {
         await using var ctx = await SeededAsync();
-        await new PrecomputedAnalyticsRefresher(ctx.Db, Service(ctx.Db), BuildService(ctx.Db), NullLogger<PrecomputedAnalyticsRefresher>.Instance)
+        await new PrecomputedAnalyticsRefresher(ctx.Db, BuildService(ctx.Db), ProService(ctx.Db), NullLogger<PrecomputedAnalyticsRefresher>.Instance)
             .RefreshMatchupsAsync(Patch, CancellationToken.None);
         var svc = Service(ctx.Db);
 
@@ -46,7 +46,7 @@ public class ChampionMatchupStatsEquivalenceTests
     public async Task Matchups_SpecificRegion_FallsBackToRawCompute()
     {
         await using var ctx = await SeededAsync();
-        await new PrecomputedAnalyticsRefresher(ctx.Db, Service(ctx.Db), BuildService(ctx.Db), NullLogger<PrecomputedAnalyticsRefresher>.Instance)
+        await new PrecomputedAnalyticsRefresher(ctx.Db, BuildService(ctx.Db), ProService(ctx.Db), NullLogger<PrecomputedAnalyticsRefresher>.Instance)
             .RefreshMatchupsAsync(Patch, CancellationToken.None);
         var svc = Service(ctx.Db);
 
@@ -57,6 +57,9 @@ public class ChampionMatchupStatsEquivalenceTests
     }
 
     private static ChampionAnalyticsComputeService Service(TranscendenceContext db) =>
+        new(db);
+
+    private static ChampionProComputeService ProService(TranscendenceContext db) =>
         new(db,
             Options.Create(new ChampionAnalyticsComputeOptions
             {
@@ -66,8 +69,7 @@ public class ChampionMatchupStatsEquivalenceTests
                 BootstrapWindowHours = 24,
                 ProvisionalWindowHours = 96,
                 MaturingWindowHours = 240
-            }),
-            NullLogger<ChampionAnalyticsComputeService>.Instance);
+            }));
 
     private static ChampionBuildComputeService BuildService(TranscendenceContext db) =>
         new(db,
